refactor: 시간 관련 함수 utils로 이동
