--- apps/web/src/components/vote-onboarding/templates/convertMinutes.ts
@@ -1,6 +1,0 @@
-export function convertMinutes(minutes: number): string {
-  const hours = Math.floor(minutes / 60);
-  const remainingMinutes = minutes % 60;
-
-  return `${hours}시간 ${remainingMinutes}분`;
-}
