--- apps/web/src/components/vote-onboarding/templates/getDeadlineInMinutes.ts
@@ -1,11 +1,0 @@
-export function getTimeDifferenceInMinutes(serverTime: string): number {
-  const serverDate = new Date(serverTime);
-
-  const currentDate = new Date();
-
-  const timeDifferenceMs = serverDate.getTime() - currentDate.getTime();
-
-  const timeDifferenceMinutes = Math.floor(timeDifferenceMs / 1000 / 60);
-
-  return timeDifferenceMinutes;
-}
