+++ apps/web/src/components/vote-voting/templates/VoteVotingLayout.tsx
@@ -10,8 +10,7 @@ import { CardList } from "../organism/CardList";
 // import { dummyPlaceList } from "@/components/vote-voting/templates/dummy";
 import { useParams } from "next/navigation";
 import { useVoteDeadline } from "@/hooks/api/useVoteDeadline";
-import { getTimeDifferenceInMinutes } from "@/components/vote-onboarding/templates/getDeadlineInMinutes";
-import { convertMinutes } from "@/components/vote-onboarding/templates/convertMinutes";
+import { convertMinutes, getTimeDifferenceInMinutes } from "@/utils/time";
 import { useVoteCandidates } from "@/hooks/api/useVoteCandidates";
 import { useVoting } from "@/hooks/api/useVoting";
 import { FixedBottomWithSpacing } from "@/components/fixed-bottom/FixedBottomWithSpacing";
